Added upgaraded VAO class, updated main shader, start creating game objects

+++ GameOpenGl/Misc/Pos.cs
@@ -8,9 +8,9 @@ namespace GameOpenGl.Misc
 {
     internal struct Pos
     {
-        int x, y;
+        float x, y;
 
-        public Pos(int x, int y)
+        public Pos(float x, float y)
         {
             this.x = x;
             this.y = y;
